backend/python/requirements.txt
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
sqlalchemy>=2.0.43
alembic>=1.14.1
psycopg[binary]>=3.2.12
pydantic>=2.11.7
httpx>=0.28.1
python-dotenv>=1.1.1
requests>=2.32.5
beautifulsoup4>=4.14.2
firebase-admin>=6.8.0
